# openapi.yaml
openapi: 3.0.3
info:
  title: Task Manager API
  version: 1.0.0
  description: |
    REST API v1 for the Task Manager app. Supports dual-mode authentication:
    Auth.js session cookies (web) or `Authorization: Bearer <accessToken>` (mobile).
    Obtain tokens via `POST /api/v1/auth/login` and refresh via `POST /api/v1/auth/refresh`.
servers:
  - url: http://localhost:3000
    description: Local development
  - url: https://projenda.vercel.app
    description: Production
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: authjs.session-token
      description: Auth.js session cookie (set after web sign-in)
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token from POST /api/v1/auth/login or /refresh
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, user]
...
    Task:
      type: object
      properties:
        id:
          type: integer
        taskId:
          type: string
          example: WEB-1
        title:
          type: string
        description:
          type: string
          nullable: true
        status:
          type: string
          enum: [backlog, todo, in_progress, qa, done, rejected, closed]
        priority:
          type: string
          enum: [low, medium, high, critical]
        projectId:
          type: integer
          nullable: true
        userId:
          type: string
        user:
          $ref: '#/components/schemas/User'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
security:
  - sessionCookie: []
  - bearerAuth: []
paths:
...
  /api/v1/tasks:
    get:
      tags: [Tasks]
      summary: List tasks
      parameters:
        - name: projectId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [backlog, todo, in_progress, qa, done, rejected, closed]
      responses:
        '200':
          description: Task list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Task'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags: [Tasks]
      summary: Create a task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title]
              properties:
                title:
                  type: string
                description:
                  type: string
                  nullable: true
                status:
                  type: string
                priority:
                  type: string
                projectId:
                  type: integer
                  nullable: true
      responses:
        '201':
          description: Task created
        '400':
          description: Validation error
        '401':
          description: Unauthorized
  /api/v1/tasks/{id}:
    get:
      tags: [Tasks]
      summary: Get a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Task details
        '404':
          description: Not found
    patch:
      tags: [Tasks]
      summary: Update a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                description:
                  type: string
                status:
                  type: string
                priority:
                  type: string
      responses:
        '200':
          description: Updated task
        '403':
          description: Forbidden
        '404':
          description: Not found
    delete:
      tags: [Tasks]
      summary: Delete a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Deleted
        '403':
          description: Forbidden
...
  /api/v1/projects/{id}:
    get:
      tags: [Projects]
      summary: Get a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Project details
    patch:
      tags: [Projects]
      summary: Update a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                description:
                  type: string
                status:
                  type: string
      responses:
        '200':
          description: Updated project